+++ git-goals/CHANGELOG.md
@@ -10,3 +10,10 @@
 
 ## [0.1.0] - Initial Release
 - Basic functionality implemented
+
+## [0.1.2] - 2024-08-12
+### Changed
+- Updated shebang in all scripts to use /usr/bin/bash
+- Improved error handling for configuration file loading
+- Enhanced help messages in main git-goals script
+- Added --help option to all subcommands